chore(toolbar): move themeLogic and dataThemeLogic to lib/logic
Continues .agents/toolbar-migration.md (Step 6). Both logics are global UI
state (dark mode, data color themes) that lib components shipped in the
toolbar depend on: HedgehogMode, JSONViewer, and the LemonColor family.
Their app-zone deps (sceneLogic, userLogic, teamLogic) already resolve to
toolbar shims, so relocating them is placement-only. themes.ts moves along
since only themeLogic imports it; dataThemeLogic's relative ./teamLogic
import becomes the absolute specifier the shim layer matches. Old paths
re-export so the ~60 scene/product importers keep working.

Graph baseline 18 -> 12 edges.
