+++ frontend/src/lib/components/CodeSnippet/CodeSnippet.tsx
@@ -13,9 +13,8 @@ import React, { useMemo, useState } from 'react'
 import { IconCollapse, IconCopy, IconExpand } from '@posthog/icons'
 
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
+import { themeLogic } from 'lib/logic/themeLogic'
 import { copyToClipboard } from 'lib/utils/copyToClipboard'
-
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
 
 import terraform from './terraformLanguage'
