+++ frontend/src/lib/components/CompactList/CompactList.tsx
@@ -6,8 +6,7 @@ import { useValues } from 'kea'
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { LemonDivider } from 'lib/lemon-ui/LemonDivider'
 import { LemonSkeleton } from 'lib/lemon-ui/LemonSkeleton'
-
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
+import { themeLogic } from 'lib/logic/themeLogic'
 
 import { EmptyMessage, EmptyMessageProps } from '../EmptyMessage/EmptyMessage'
