+++ frontend/src/lib/components/HedgehogMode/HedgehogMode.tsx
@@ -1,10 +1,10 @@
 import { useActions, useValues } from 'kea'
 import { Suspense } from 'react'
 
+import { themeLogic } from 'lib/logic/themeLogic'
 import { inStorybook } from 'lib/utils/dom'
 import { lazyWithRetry } from 'lib/utils/retryImport'
 
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
 import { toolbarConfigLogic } from '~/toolbar/toolbarConfigLogic'
 
 import { useShortcut } from '../Shortcuts/useShortcut'
