+++ frontend/src/lib/components/JSONViewer.tsx
@@ -4,10 +4,9 @@ import type { ReactJsonViewProps } from '@microlink/react-json-view'
 import { useValues } from 'kea'
 import { Suspense } from 'react'
 
+import { themeLogic } from 'lib/logic/themeLogic'
 import { WrappingLoadingSkeleton } from 'lib/ui/WrappingLoadingSkeleton/WrappingLoadingSkeleton'
 import { lazyWithRetry } from 'lib/utils/retryImport'
-
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
 
 const ReactJson = lazyWithRetry(() => import('@microlink/react-json-view'))
