+++ frontend/src/lib/components/SeriesGlyph.tsx
@@ -1,10 +1,9 @@
 import { useValues } from 'kea'
 
 import { getSeriesColor } from 'lib/colors'
+import { themeLogic } from 'lib/logic/themeLogic'
 import { RGBToRGBA, hexToRGBA, lightenDarkenColor } from 'lib/utils/colors'
 import { alphabet } from 'lib/utils/strings'
-
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
 
 interface SeriesGlyphProps {
     className?: string
