+++ frontend/src/lib/components/Sharing/SharingModal.test.tsx
@@ -3,9 +3,9 @@ import '@testing-library/jest-dom'
 import { cleanup, render, screen, within } from '@testing-library/react'
 import { expectLogic } from 'kea-test-utils'
 
+import { themeLogic } from 'lib/logic/themeLogic'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
 import { useAvailableFeatures } from '~/mocks/features'
 import { useMocks } from '~/mocks/jest'
 import { NodeKind } from '~/queries/schema/schema-general'
