+++ frontend/src/lib/lemon-ui/LemonMarkdown/MermaidDiagram.test.tsx
@@ -6,7 +6,7 @@ jest.mock('kea', () => ({
     useValues: () => ({ isDarkModeOn: false }),
 }))
 
-jest.mock('~/layout/navigation-3000/themeLogic', () => ({
+jest.mock('lib/logic/themeLogic', () => ({
     themeLogic: { values: {} },
 }))
