+++ frontend/src/lib/lemon-ui/LemonMarkdown/MermaidDiagram.tsx
@@ -3,8 +3,7 @@ import { useEffect, useId, useState } from 'react'
 
 import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 import { Spinner } from 'lib/lemon-ui/Spinner'
-
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
+import { themeLogic } from 'lib/logic/themeLogic'
 
 type MermaidApi = {
     initialize: (config: Record<string, unknown>) => void
