+++ frontend/src/lib/ui/Colors/Colors.stories.tsx
@@ -3,8 +3,7 @@ import { useValues } from 'kea'
 import { useEffect, useState } from 'react'
 
 import { LemonSlider } from 'lib/lemon-ui/LemonSlider'
-
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
+import { themeLogic } from 'lib/logic/themeLogic'
 
 const meta: Meta = {
     title: 'UI/Colors',
